update docs + example
