+++ framework/examples/myproject/smoke_logs_test.go
@@ -1,7 +1,10 @@
 package examples
 
 import (
+	"bufio"
 	"fmt"
+	"io"
+	"regexp"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -22,23 +25,38 @@ func TestLogsSmoke(t *testing.T) {
 	in, err := framework.Load[CfgLogs](t)
 	require.NoError(t, err)
 	// most simple checks, save all the logs and check (CRIT|PANIC|FATAL) log levels
-	//t.Cleanup(func() {
-	//	err := framework.SaveAndCheckLogs(t)
-	//	require.NoError(t, err)
-	//})
 	t.Cleanup(func() {
-		// save all the logs to default directory "logs/docker-$test_name"
-		logs, err := framework.SaveContainerLogs(fmt.Sprintf("%s-%s", framework.DefaultCTFLogsDir, t.Name()))
+		err := framework.SaveAndCheckLogs(t)
 		require.NoError(t, err)
-		// check that CL nodes has no errors (CRIT|PANIC|FATAL) levels
-		err = framework.CheckCLNodeContainerErrors()
+	})
+
+	re := regexp.MustCompile(`name=HeadReporter version=\d+`)
+	t.Cleanup(func() {
+		err := framework.StreamCTFContainerLogsFanout(
+			framework.LogStreamConsumer{
+				Name: "custom-regex-assert",
+				Consume: func(logStreams map[string]io.ReadCloser) error {
+					for name, stream := range logStreams {
+						scanner := bufio.NewScanner(stream)
+						found := false
+						for scanner.Scan() {
+							if re.MatchString(scanner.Text()) {
+								found = true
+								break
+							}
+						}
+						if err := scanner.Err(); err != nil {
+							return fmt.Errorf("scan %s: %w", name, err)
+						}
+						if !found {
+							return fmt.Errorf("missing HeadReporter log in %s", name)
+						}
+					}
+					return nil
+				},
+			},
+		)
 		require.NoError(t, err)
-		// do custom assertions
-		for _, l := range logs {
-			matches, err := framework.SearchLogFile(l, " name=HeadReporter version=\\d")
-			require.NoError(t, err)
-			_ = matches
-		}
 	})
 
 	bc, err := blockchain.NewBlockchainNetwork(in.BlockchainA)
